CHANNELMGR-1849 hide menu options for authors that will never be enabled for them
Updated the unit test for the new visability options of the menu as a whole. Also described the menu for authors.

+++ channel-manager/frontend-ng/src/app/channel/page/actions/pageActions.service.js
@@ -66,34 +66,36 @@ class PageActionsService extends MenuService {
       });
     }
 
-    menu
-      .addAction('properties', {
-        translationKey: 'TOOLBAR_MENU_PAGE_PROPERTIES',
-        isEnabled: () => this._canEditPage(),
-        onClick: () => this._pageProperties(),
-      })
-      .addDivider()
-      .addAction('copy', {
-        translationKey: 'TOOLBAR_MENU_PAGE_COPY',
-        isEnabled: () => this._canCopyPage(),
-        onClick: () => this._copyPage(),
-      })
-      .addAction('move', {
-        translationKey: 'TOOLBAR_MENU_PAGE_MOVE',
-        isEnabled: () => this._canEditPage(),
-        onClick: () => this._movePage(),
-      })
-      .addAction('delete', {
-        translationKey: 'TOOLBAR_MENU_PAGE_DELETE',
-        isEnabled: () => this._canEditPage(),
-        onClick: () => this._deletePage(),
-      })
-      .addDivider()
-      .addAction('new', {
-        translationKey: 'TOOLBAR_MENU_PAGE_NEW',
-        isEnabled: () => this._canAddNewPage(),
-        onClick: () => this._newPage(),
-      });
+    if (this._hasWriteAccess()) {
+      menu
+        .addAction('properties', {
+          translationKey: 'TOOLBAR_MENU_PAGE_PROPERTIES',
+          isEnabled: () => this._canEditPage(),
+          onClick: () => this._pageProperties(),
+        })
+        .addDivider()
+        .addAction('copy', {
+          translationKey: 'TOOLBAR_MENU_PAGE_COPY',
+          isEnabled: () => this._canCopyPage(),
+          onClick: () => this._copyPage(),
+        })
+        .addAction('move', {
+          translationKey: 'TOOLBAR_MENU_PAGE_MOVE',
+          isEnabled: () => this._canEditPage(),
+          onClick: () => this._movePage(),
+        })
+        .addAction('delete', {
+          translationKey: 'TOOLBAR_MENU_PAGE_DELETE',
+          isEnabled: () => this._canEditPage(),
+          onClick: () => this._deletePage(),
+        })
+        .addDivider()
+        .addAction('new', {
+          translationKey: 'TOOLBAR_MENU_PAGE_NEW',
+          isEnabled: () => this._canAddNewPage(),
+          onClick: () => this._newPage(),
+        });
+    }
   }
 
   onOpenMenu() {
@@ -123,6 +125,10 @@ class PageActionsService extends MenuService {
 
   _hasPageExtensions() {
     return this.ExtensionService.hasExtensions('page');
+  }
+
+  _hasWriteAccess() {
+    return this.SessionService.hasWriteAccess();
   }
 
   _canAccessAnItem() {
